requirements.txt
alembic==1.10.2
click==8.1.3
Flask==2.2.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
gunicorn==20.1.0
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
python-decouple==3.8
python-dotenv==0.21.1
SQLAlchemy==1.4.46
typing_extensions==4.5.0
Werkzeug==2.2.2
zipp==3.11.0
